# docs/openapi.yaml
openapi: 3.0.3
info:
  title: G-Flow Super-App API
  description: |
    REST API for G-Flow super-app ecosystem (Wallet, Ride, Food, Send, Admin).
    Synchronized with docs/API CONTRACT.txt v1.0-FINAL.
    JWT authentication (Bearer token). Idempotency via X-Idempotency-Key header.
    Response envelope: `{ success, data, meta }`. Errors: `{ success, error }`.
  version: 1.0.0
  contact:
    name: G-Flow Engineering
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.g-flow.com
    description: Production
security:
  - bearerAuth: []
tags:
  - name: Auth
    description: Authentication & user lifecycle
  - name: Wallet
    description: PayPulse wallet, top-up, transfer, withdrawal, ledger
  - name: Ride
    description: G-Ride ride-hailing
  - name: Food
    description: G-Food food delivery
  - name: Send
    description: G-Send instant logistics (multi-stop)
  - name: Admin
    description: Admin operations (dashboard, users, reversal)
  - name: Health
    description: Health & readiness
paths:
  /health:
    get:
      tags: [Health]
      summary: Liveness check
      responses:
        '200':
          description: Service alive
  /healthz/liveness:
    get:
      tags: [Health]
      summary: Liveness probe (HTTP only)
      responses:
        '200':
          description: Alive
  /healthz/readiness:
    get:
      tags: [Health]
      summary: Readiness probe (DB + Redis)
      responses:
        '200':
          description: Ready
        '503':
          description: Not ready
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, email, phone, password, user_type]
              properties:
                full_name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string, minLength: 8 }
                user_type:
                  type: string
                  enum: [customer, driver, merchant]
                vehicle_type:
                  type: string
                  enum: [motorcycle, car, van]
                license_number: { type: string }
                license_expiry: { type: string, format: date }
                vehicle_plate: { type: string }
      responses:
        '201':
          description: User created
        '409':
          description: Email/phone already exists
        '422':
          description: Validation error
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                phone: { type: string }
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  user_id: { type: string, format: uuid }
                  user_type: { type: string }
        '401':
          description: Invalid credentials
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke current token
      responses:
        '200':
          description: Logged out
  /api/v1/auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New tokens issued
  /api/v1/wallets/{wallet_id}/topup:
    post:
      tags: [Wallet]
      summary: Top-up a wallet (idempotent)
      parameters:
        - { name: wallet_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, idempotency_key]
              properties:
                amount: { type: integer, minimum: 1 }
                idempotency_key: { type: string }
      responses:
        '201':
          description: Top-up success
        '422':
          description: Business validation error
  /api/v1/wallets/{wallet_id}/transfer:
    post:
      tags: [Wallet]
      summary: Transfer between wallets (idempotent)
      parameters:
        - { name: wallet_id, in: path, required: true, schema: { type: string, format: uuid }, description: Source wallet }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_wallet_id, amount]
              properties:
                to_wallet_id: { type: string, format: uuid }
                amount: { type: integer, minimum: 1 }
                description: { type: string }
                reference_type: { type: string }
                idempotency_key: { type: string }
      responses:
        '200':
          description: Transfer success
        '409':
          description: Conflict or duplicate idempotency key
        '422':
          description: Insufficient balance / self-transfer / limit
  /api/v1/wallets/{wallet_id}/balance:
    get:
      tags: [Wallet]
      summary: Get wallet balance
      parameters:
        - { name: wallet_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Balance details
  /api/v1/wallets/{wallet_id}/withdrawal:
    post:
      tags: [Wallet]
      summary: Request a withdrawal (driver/merchant only)
      parameters:
        - { name: wallet_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: integer }
                bank_account_number: { type: string }
                bank_name: { type: string }
      responses:
        '201':
          description: Withdrawal pending approval
        '422':
          description: Validation error
  /api/v1/ledger/audit:
    get:
      tags: [Wallet, Admin]
      summary: Ledger audit trail (admin)
      parameters:
        - { name: start_date, in: query, schema: { type: string, format: date } }
        - { name: end_date, in: query, schema: { type: string, format: date } }
        - { name: wallet_id, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        '200':
          description: Ledger entries with balance verification
  /api/v1/rides/book:
    post:
      tags: [Ride]
      summary: Book a ride (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickup_lat, pickup_lng, dropoff_lat, dropoff_lng]
              properties:
                pickup_lat: { type: number }
                pickup_lng: { type: number }
                dropoff_lat: { type: number }
                dropoff_lng: { type: number }
                preferred_vehicle: { type: string, enum: [REGULAR, PREMIUM] }
      responses:
        '201':
          description: Ride order created
        '422':
          description: Insufficient escrow / invalid coordinates
  /api/v1/rides/{order_id}/accept:
    post:
      tags: [Ride]
      summary: Driver accepts an order
      parameters:
        - { name: order_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driver_id]
              properties:
                driver_id: { type: string, format: uuid }
      responses:
        '200':
          description: Order assigned to driver
        '409':
          description: Already assigned (race condition)
        '422':
          description: Driver capacity / balance check
  /api/v1/rides/{order_id}/status:
    patch:
      tags: [Ride]
      summary: Update ride status
      parameters:
        - { name: order_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [DRIVER_ARRIVED, TRIP_STARTED, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Status updated
        '409':
          description: Invalid state transition
  /api/v1/rides/{order_id}:
    get:
      tags: [Ride]
      summary: Get ride details
      parameters:
        - { name: order_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Ride details
  /api/v1/rides/available:
    get:
      tags: [Ride]
      summary: List available ride orders (driver)
      responses:
        '200':
          description: Available orders
  /api/v1/merchants:
    get:
      tags: [Food]
      summary: List merchants
      security: []
      parameters:
        - { name: lat, in: query, schema: { type: number } }
        - { name: lng, in: query, schema: { type: number } }
        - { name: radius_km, in: query, schema: { type: number, default: 10 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200':
          description: Merchant list
  /api/v1/merchants/{merchant_id}/menu:
    get:
      tags: [Food]
      summary: Get merchant menu with option groups
      security: []
      parameters:
        - { name: merchant_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Menu structure
  /api/v1/food-orders:
    post:
      tags: [Food]
      summary: Create a food order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [merchant_id, items]
              properties:
                merchant_id: { type: string, format: uuid }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id: { type: string, format: uuid }
                      quantity: { type: integer }
                      selected_options:
                        type: array
                        items:
                          type: object
                delivery_address: { type: string }
                delivery_lat: { type: number }
                delivery_lng: { type: number }
                payment_method: { type: string, enum: [WALLET, CASH] }
      responses:
        '201':
          description: Food order created
  /api/v1/food-orders/{order_id}:
    get:
      tags: [Food]
      summary: Get food order details
      parameters:
        - { name: order_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Food order details
  /api/v1/food-orders/available:
    get:
      tags: [Food]
      summary: List available food orders (driver)
      responses:
        '200':
          description: Available orders
  /api/v1/send-orders:
    post:
      tags: [Send]
      summary: Create a send order (multi-stop)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [package_name, stops]
              properties:
                package_name: { type: string }
                weight: { type: number }
                package_type: { type: string }
                stops:
                  type: array
                  items:
                    type: object
                    properties:
                      recipient_name: { type: string }
                      address: { type: string }
                      lat: { type: number }
                      lng: { type: number }
                payment_method: { type: string, enum: [WALLET, CASH] }
      responses:
        '201':
          description: Send order created
  /api/v1/send-orders/{order_id}:
    get:
      tags: [Send]
      summary: Get send order details with stops
      parameters:
        - { name: order_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Send order details
  /api/v1/send-orders/available:
    get:
      tags: [Send]
      summary: List available send orders (driver)
      responses:
        '200':
          description: Available orders
  /api/v1/drivers/{driver_id}/orders:
    get:
      tags: [Ride, Food, Send]
      summary: List active orders for a driver
      parameters:
        - { name: driver_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Driver active orders
  /api/v1/admin/dashboard/kpis:
    get:
      tags: [Admin]
      summary: Admin dashboard KPIs
      responses:
        '200':
          description: KPI values
  /api/v1/admin/dashboard/transactions:
    get:
      tags: [Admin]
      summary: Latest transactions (admin)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Transaction list
  /api/v1/admin/ledger:
    get:
      tags: [Admin]
      summary: Ledger audit trail (admin)
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: date_from, in: query, schema: { type: string } }
        - { name: date_to, in: query, schema: { type: string } }
        - { name: wallet_type, in: query, schema: { type: string } }
      responses:
        '200':
          description: Ledger entries
  /api/v1/admin/ledger/verify/{wallet_id}:
    get:
      tags: [Admin]
      summary: Verify wallet balance
      parameters:
        - { name: wallet_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Balance verification result
  /api/v1/admin/ledger/export:
    post:
      tags: [Admin]
      summary: Export ledger (CSV/JSON)
      responses:
        '200':
          description: Export file
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users (admin)
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: role, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: User list
  /api/v1/admin/users/{id}/freeze:
    patch:
      tags: [Admin]
      summary: Freeze a user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User frozen
  /api/v1/admin/users/{id}/suspend:
    patch:
      tags: [Admin]
      summary: Suspend a user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User suspended
  /api/v1/admin/users/{id}/ban:
    patch:
      tags: [Admin]
      summary: Ban a user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: User banned
  /api/v1/admin/transactions/{id}/reverse:
    post:
      tags: [Admin]
      summary: Reverse a transaction (2FA required)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
                notes: { type: string }
      responses:
        '200':
          description: Transaction reversed
        '401':
          description: Invalid 2FA token
        '429':
          description: Lockout (3 failed attempts)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    admin2FA:
      type: apiKey
      in: header
      name: X-Admin-2FA-Token
      description: TOTP token required for admin reversal